define first roles in the system,extract data from register request and save user using repos considering user role

+++ src/main/java/com/lukasz/project/repository/AdminRepository.java
@@ -1,13 +1,13 @@
 package com.lukasz.project.repository;
 
-import com.lukasz.project.entity.Admin;
+import com.lukasz.project.model.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface AdminRepository extends JpaRepository<Admin, Long> {
+public interface AdminRepository extends JpaRepository<Admin, Integer> {
     Optional<Admin> findByEmail(String email);
     boolean existsByEmail(String email);
 }
