+++ src/main/java/com/lukasz/project/repository/CompanyRepository.java
@@ -1,11 +1,11 @@
 package com.lukasz.project.repository;
 
-import com.lukasz.project.entity.Company;
+import com.lukasz.project.model.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CompanyRepository extends JpaRepository<Company, Long> {
+public interface CompanyRepository extends JpaRepository<Company, Integer> {
 
 
 }
