+++ src/main/java/com/lukasz/project/repository/OfferRepository.java
@@ -1,9 +1,9 @@
 package com.lukasz.project.repository;
 
-import com.lukasz.project.entity.Offer;
+import com.lukasz.project.model.Offer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OfferRepository extends JpaRepository<Offer,Long> {
+public interface OfferRepository extends JpaRepository<Offer,Integer> {
 }
