+++ src/main/java/com/lukasz/project/service/CompanyService.java
@@ -1,6 +1,6 @@
 package com.lukasz.project.service;
 
-import com.lukasz.project.entity.Company;
+import com.lukasz.project.model.Company;
 import com.lukasz.project.repository.CompanyRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,7 +14,7 @@ public class CompanyService {
     public void create(Company company){
         companyRepository.save(company);
     }
-    public void delete(Long id){
+    public void delete(Integer id){
         companyRepository.deleteById(id);
     }
 }
